+++ app/process.js
@@ -1,0 +1,6 @@
+const {ipcRenderer} = require('electron')
+
+ipcRenderer.on('playSound', (event, data) => {
+  let audio = new Audio(`audio/${data}.wav`)
+  audio.play()
+})